requirements.txt
Django==5.2.4
channels==4.3.1
channels-redis==4.3.0
daphne==4.2.1
django-bootstrap-icons==0.9.0
Pillow==11.3.0
django-imagekit==5.0.0
djangorestframework==3.16.0
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0
Babel==2.17.0
django-cors-headers==4.7.0
redis==6.4.0
asgiref==3.9.1
whitenoise==6.10.0
websockets==15.0.1
apscheduler==3.10.4
psycopg[binary]==3.3.1
python-dotenv==1.0.1
dj-database-url==2.2.0
supabase==2.10.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
